This commit must never be merged!! Comments and sample config for the tool-database branch
I have deleted all the sample configs that don't work with this branch.
I get frustrated when trying an experimental branch and it is not easy
to see which config is actually meant towork. No such ambiguity here, as
I deleted all the other configs. But merging that would be bad.
This branch is pushed purely for fiddling with and discussion. It is not
suitable for actual hardware.
The document ABOUT_THIS_BRANCH contains a summary and is meant as a scratch-
pad for comments.


Signed-off-by: Andy Pugh <andy@bodgesoc.org>