deb: new debian/configure and debian/control*
This one's written in python instead of bash, and has a cleaner internal
structure.

This handles the flavored debs of the unified-build-candidate branch.

The new debian packaging produces linuxcnc.deb, linuxcnc-dev.deb,
and linuxcnc-doc-*.deb, just like before, and in addition produces
linuxcnc-$FLAVOR.deb for each flavor enabled by debian/configure.
The available flavors are:

    posix (what we used to call sim)

    rtai-kernel (what we used to call rt)

    rt-preempt

    xenomai (FIXME: not fully implemented yet)

    xenomai-kernel (FIXME: not fully implemented yet)

linuxcnc.deb depends on at least one of the flavor debs, and suggests
all of them (to make them easy to find in the package manager, mostly).