deb/conf: accept -a to mean -r
"debian/configure -r" is supposed to mean "configure for the running
realtime kernel", and "debian/configure -a" is supposed to mean "configure
for any realtime headers you can find on the system".

The buildbot ill-advisedly runs 'debian/configure -a', even though it
knows it's going to run runtests later (so the kernel it builds for must
be the one that's running).  So really -r would be more correct.

In practice it's fine because all the buildslave VMs only have one
realtime kernel installed, so there's no difference between -a and -r
on those hosts.

If someone cares strongly about the difference between the old "-a"
and the new "-a" we can reinstate it.

As an aside, it probably makes sense to add a "debian/configure --all"
option that examines the apt database and makes the dsc build-depend on
all *available* realtime kernel headers.