add a linuxcnc-test.deb, with all the tests in it
This commit is not ready yet.  FIXME:

    runtests gets installed gzipped and without execute permissions

    the runtests script tries to run scripts/rip-environment first thing :-/

The ubc branch has problems that our tests missed because we don't follow
the 'test like you fly' principle.

There are differences between run-in-place builds and installed builds,
and our current testing method only tests run-in-place, so we can miss
problems with installed configurations.  This commit adds a deb with
just our tests in it, so we can test installed programs too.