linuxcnc-test.deb: don't compress or clear execute permissions on the tests
FIXME: another problem with linuxcnc-test.deb is that the user doesn't
    have write permission to /usr/share/doc/linuxcnc/tests, so all the
    tests fail

With this commit, if i cp the tests directory to my home dir and edit
runtests to not try to source rip-environment, the tests start running.
Some tests pass.  Some still fail, and the runtests script itself falls
over part way into the test list.